+++ apps/frontend/localization/ru.json
@@ -121,6 +121,13 @@
       "delete": "Удалить",
       "details": "Детали"
     },
+    "task_edit": {
+      "title": "Редактирование задачи",
+      "heading": "Редактирование Задачи",
+      "description": "Редактирование задачи.",
+      "task": "Задача",
+      "Delete": "Удалить"
+    },
     "account": {
       "title": "Учётная запись",
       "heading": "Учётная Запись",
@@ -245,14 +252,6 @@
     },
     "edit": "Редактировать",
     "editing": "Редактируем..."
-  },
-
-  "task_edit": {
-    "title": "Редактирование задачи",
-    "heading": "Редактирование Задачи",
-    "description": "Редактирование задачи.",
-    "task": "Задача",
-    "Delete": "Удалить"
   },
   "place": {
     "title": "Заглавие",
